+++ packages/polaris-viz-core/src/index.ts
@@ -92,6 +92,9 @@ export {
   getColorVisionEventAttrs,
   getColorVisionStylesForActiveIndex,
   getRoundedRectPath,
+  changeColorOpacity,
+  changeGradientOpacity,
+  getAverageColor,
 } from './utilities';
 export {
   useSparkBar,