+++ packages/polaris-viz-native/src/index.ts
@@ -1,1 +1,18 @@
 export {SparkLineChart, PolarisVizProvider, SparkBarChart} from './components';
+
+export {
+  createGradient,
+  curveStepRounded,
+  getFilteredSeries,
+  getSeriesColors,
+  isGradientType,
+  uniqueId,
+  createTheme,
+  getColorVisionStylesForActiveIndex,
+  getColorVisionEventAttrs,
+  changeColorOpacity,
+  changeGradientOpacity,
+  getAverageColor,
+  paddingStringToObject,
+  removeFalsyValues,
+} from '@shopify/polaris-viz-core';