+++ packages/polaris-viz/src/components/Docs/stories/components/UtilitiesHeader/index.ts
@@ -1,0 +1,1 @@
+export {UtilitiesHeader} from './UtilitiesHeader';