+++ packages/polaris-viz/src/components/LineChart/LineChart.tsx
@@ -5,7 +5,12 @@ import type {
   XAxisOptions,
   YAxisOptions,
 } from '@shopify/polaris-viz-core';
-import {isGradientType, uniqueId} from '@shopify/polaris-viz-core';
+import {
+  isGradientType,
+  uniqueId,
+  changeColorOpacity,
+  getAverageColor,
+} from '@shopify/polaris-viz-core';
 
 import {formatTooltipDataForLinearCharts} from '../../utilities/formatTooltipDataForLinearCharts';
 import type {RenderTooltipContentData} from '../../types';
@@ -13,8 +18,6 @@ import {TooltipContent} from '../../components/TooltipContent';
 import {ChartContainer} from '../../components/ChartContainer';
 import {useThemeSeriesColors} from '../../hooks/useThemeSeriesColors';
 import {
-  changeColorOpacity,
-  getAverageColor,
   getXAxisOptionsWithDefaults,
   getYAxisOptionsWithDefaults,
 } from '../../utilities';